+++ test/mocha/helpers.js
@@ -1,20 +1,20 @@
 /*!
  * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const bedrock = require('bedrock');
-const brAccount = require('bedrock-account');
-const {getAppIdentity} = require('bedrock-app-identity');
-const brPassport = require('bedrock-passport');
-const database = require('bedrock-mongodb');
+import * as bedrock from '@bedrock/core';
+import * as brAccount from '@bedrock/account';
+import * as brPassport from '@bedrock/passport';
+import * as database from '@bedrock/mongodb';
+import {agent} from '@bedrock/https-agent';
+import {createRequire} from 'module';
+import {getAppIdentity} from '@bedrock/app-identity';
+import {mockData} from './mock.data.js';
+import sinon from 'sinon';
+const require = createRequire(import.meta.url);
 const {Ed25519Signature2020} = require('@digitalbazaar/ed25519-signature-2020');
-const {agent} = require('bedrock-https-agent');
-const sinon = require('sinon');
 const {ZcapClient} = require('@digitalbazaar/ezcap');
-const mockData = require('./mock.data');
 
-exports.createMeter = async ({capabilityAgent, type}) => {
+export async function createMeter({capabilityAgent, type}) {
   if(!(type && capabilityAgent)) {
     throw new Error(`"capabilityAgent" and "type" must be provided.`);
   }
@@ -41,9 +41,9 @@ exports.createMeter = async ({capabilityAgent, type}) => {
   // return usage capability
   const {id} = meter;
   return {id: `${meterService}/${id}`};
-};
+}
 
-exports.stubPassport = async ({email = 'alpha@example.com'} = {}) => {
+export function stubPassport({email = 'alpha@example.com'} = {}) {
   const passportStub = sinon.stub(brPassport, 'optionallyAuthenticated');
   passportStub.callsFake((req, res, next) => {
     req.user = {
@@ -52,26 +52,27 @@ exports.stubPassport = async ({email = 'alpha@example.com'} = {}) => {
     next();
   });
   return passportStub;
-};
+}
 
-exports.prepareDatabase = async mockData => {
-  await exports.removeCollections();
+export async function prepareDatabase(mockData) {
+  await removeCollections();
   await insertTestData(mockData);
-};
+}
 
-exports.removeCollections = async (
+export async function removeCollections(
   collectionNames = [
     'account',
     'profile-profileAgent'
-  ]) => {
+  ]) {
   await database.openCollections(collectionNames);
   for(const collectionName of collectionNames) {
     await database.collections[collectionName].deleteMany({});
   }
-};
+}
 
-exports.removeCollection =
-  async collectionName => exports.removeCollections([collectionName]);
+export async function removeCollection(collectionName) {
+  return removeCollections([collectionName]);
+}
 
 async function insertTestData(mockData) {
   const records = Object.values(mockData.accounts);