+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # bedrock-profile ChangeLog
 
+## 16.0.0 - 2022-04-xx
+
+### Changed
+- **BREAKING**: Rename package to `@bedrock/profile`.
+- **BREAKING**: Convert to module (ESM).
+- **BREAKING**: Remove default export.
+- **BREAKING**: Require node 14.x.
+
 ## 15.1.0 - 2022-03-29
 
 ### Changed